% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApplication.R
\name{startApplicationLocal}
\alias{startApplicationLocal}
\title{Start Application}
\usage{
startApplicationLocal(...)
}
\arguments{
\item{...}{passed to \link[shiny]{runApp}


#' @export
#' @rdname startApplication
startApplication <- function() {
  packageDir <- system.file(package = "EcomAnalytics")
  shiny::runApp(
    paste0(packageDir, "/app"),
    port = 3838,
    host = "0.0.0.0"
}}
}
\description{
Start Application
}
