% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpiCalculations.R, R/plotOutputs.R,
%   R/tableOutputs.R
\name{calcRevenueShop}
\alias{calcRevenueShop}
\alias{calcCustomersShop}
\alias{calcNumProdsShop}
\alias{calcRevenueI}
\alias{calcQuantileI}
\alias{calcNumProdsI}
\alias{trendDist}
\alias{timeDist}
\alias{trendDistI}
\alias{timeDistI}
\alias{calcTopProductsShop}
\alias{emptyProdTable}
\alias{calcLowProductsShop}
\alias{calcTopProductsI}
\alias{calcLowProductsI}
\title{Data Preparation Functions}
\usage{
calcRevenueShop(ecomData)

calcCustomersShop(ecomData)

calcNumProdsShop(ecomData)

calcRevenueI(ecomData, customerID)

calcQuantileI(ecomData, customerID)

calcNumProdsI(ecomData, customerID)

trendDist(ecomData, dateSpan, trendVar)

timeDist(ecomData, timeVar)

trendDistI(ecomData, customerID, dateSpan, trendVarI)

timeDistI(ecomData, customerID, timeVarI)

calcTopProductsShop(ecomData, numProducts, dateSpan)

emptyProdTable()

calcLowProductsShop(ecomData, numProducts, dateSpan)

calcTopProductsI(ecomData, customerID, numProducts, dateSpan)

calcLowProductsI(ecomData, customerID, numProducts, dateSpan)
}
\arguments{
\item{ecomData}{data object}
}
\description{
Data Preparation Functions

Plot Output Functions
}
