% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{box}
\alias{box}
\alias{panelTitle}
\alias{panelSelectInput}
\alias{shopLevelKpis}
\alias{shopLevelProductRanking}
\alias{shopLevelTimeAnalysis}
\alias{shopLevelTrendAnalysis}
\alias{customerIdSelection}
\alias{iLevelKpis}
\alias{iLevelProductRanking}
\alias{iLevelTimeAnalysis}
\alias{iLevelTrendAnalysis}
\title{UI Elements}
\usage{
box(...)

panelTitle(sideBarWidth)

panelSelectInput(buttonWidth)

shopLevelKpis()

shopLevelProductRanking()

shopLevelTimeAnalysis()

shopLevelTrendAnalysis()

customerIdSelection()

iLevelKpis()

iLevelProductRanking()

iLevelTimeAnalysis()

iLevelTrendAnalysis()
}
\arguments{
\item{...}{UI elements for the box}

\item{sideBarWidth}{(numeric) width of sidebar}

\item{buttonWidth}{button width}
}
\description{
UI Elements
}
